+++ app/vmui/packages/vmui/src/components/Views/GroupView/GroupLogsFieldRow.tsx
@@ -2,14 +2,17 @@ import { FC, memo } from "preact/compat";
 import classNames from "classnames";
 import useDeviceDetect from "../../../hooks/useDeviceDetect";
 import FieldRowActions from "./GroupLogsFieldRowMenu/FieldRowActions";
+import { FieldStreamIcon } from "../../Main/Icons";
+import Tooltip from "../../Main/Tooltip/Tooltip";
 
 interface Props {
   field: string;
   value: string;
+  isStreamField: boolean;
   hideGroupButton?: boolean;
 }
 
-const GroupLogsFieldRow: FC<Props> = ({ field, value, hideGroupButton = false }) => {
+const GroupLogsFieldRow: FC<Props> = ({ field, value, isStreamField, hideGroupButton = false }) => {
   const { isMobile } = useDeviceDetect();
 
   return (
@@ -19,12 +22,18 @@ const GroupLogsFieldRow: FC<Props> = ({ field, value, hideGroupButton = false })
       "vm-group-logs-row-fields-item_mobile": isMobile
     })}
     >
+      <td className="vm-group-logs-row-fields-item__icon">
+        <Tooltip title="Stream field">
+          {isStreamField ? <FieldStreamIcon/> : null}
+        </Tooltip>
+      </td>
       <td className="vm-group-logs-row-fields-item__key">{field}</td>
       <td className="vm-group-logs-row-fields-item__value">{value}</td>
       <td className="vm-group-logs-row-fields-item-controls">
         <FieldRowActions
           field={field}
           value={value}
+          isStreamField={isStreamField}
           hideGroupButton={hideGroupButton}
         />
       </td>