+++ app/vmui/packages/vmui/src/components/Views/GroupView/GroupLogsFieldRowMenu/FieldRowActions.tsx
@@ -12,10 +12,11 @@ import Popper from "../../../Main/Popper/Popper";
 type Props = {
   field: string;
   value: string;
+  isStreamField: boolean;
   hideGroupButton: boolean;
 }
 
-const FieldRowActions: FC<Props> = ({ field, value, hideGroupButton }) => {
+const FieldRowActions: FC<Props> = ({ field, value, isStreamField, hideGroupButton }) => {
   const {
     value: openContextMenu,
     setFalse: handleCloseContextMenu,
@@ -40,6 +41,7 @@ const FieldRowActions: FC<Props> = ({ field, value, hideGroupButton }) => {
       <FieldFilterInclude
         field={field}
         value={value}
+        isStreamField={isStreamField}
       />
 
       <div ref={buttonRef}>
@@ -62,6 +64,7 @@ const FieldRowActions: FC<Props> = ({ field, value, hideGroupButton }) => {
             <FieldFilterExclude
               field={field}
               value={value}
+              isStreamField={isStreamField}
               onClose={handleCloseContextMenu}
             />
             <FieldMessageToggle